requirements.txt
torch>=1.2.0
torchvision==0.9.1
matplotlib>=2.1.1
numpy>=1.13.3
pytorch-lightning==1.3.8
continuum==1.2.1
hydra-core==1.0.6
test_tube>=0.7.5
wandb>=0.10.22
opencv-python==4.4.0.46
shortuuid==1.0.1
plotly
einops==0.3.2
kaleido==0.1.0
segmentation_models_pytorch>=0.2.1
setuptools==59.5.0
torchmetrics==0.6.0